Rename Literal dir to Literals To make it consistant
