+++ src/Ast/Literals/BooleanLiteral.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Serhii\GoodbyeHtml\Ast\Literal;
+namespace Serhii\GoodbyeHtml\Ast\Literals;
 
 use Serhii\GoodbyeHtml\Ast\Expression;
 use Serhii\GoodbyeHtml\Token\Token;
