+++ src/Ast/Literals/IntegerLiteral.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Serhii\GoodbyeHtml\Ast\Literal;
+namespace Serhii\GoodbyeHtml\Ast\Literals;
 
 use Serhii\GoodbyeHtml\Ast\Expression;
 use Serhii\GoodbyeHtml\Token\Token;
 
-readonly class FloatLiteral implements Expression
+readonly class IntegerLiteral implements Expression
 {
-    public function __construct(public Token $token, public float $value)
+    public function __construct(public Token $token, public int $value)
     {
     }
