+++ src/Ast/Literals/StringLiteral.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Serhii\GoodbyeHtml\Ast\Literal;
+namespace Serhii\GoodbyeHtml\Ast\Literals;
 
 use Serhii\GoodbyeHtml\Ast\Expression;
 use Serhii\GoodbyeHtml\Token\Token;
 
-readonly class NullLiteral implements Expression
+readonly class StringLiteral implements Expression
 {
-    public function __construct(public Token $token)
+    public function __construct(public Token $token, public string $value)
     {
     }
 
@@ -20,6 +20,6 @@ readonly class NullLiteral implements Expression
 
     public function string(): string
     {
-        return '';
+        return '"' . $this->value . '"';
     }
 }
