+++ src/CoreParser/CoreParser.php
@@ -6,20 +6,20 @@ namespace Serhii\GoodbyeHtml\CoreParser;
 
 use Closure;
 use Serhii\GoodbyeHtml\Ast\Statements\BlockStatement;
-use Serhii\GoodbyeHtml\Ast\Literal\BooleanLiteral;
+use Serhii\GoodbyeHtml\Ast\Literals\BooleanLiteral;
 use Serhii\GoodbyeHtml\Ast\Expression;
 use Serhii\GoodbyeHtml\Ast\Statements\ExpressionStatement;
-use Serhii\GoodbyeHtml\Ast\Literal\FloatLiteral;
+use Serhii\GoodbyeHtml\Ast\Literals\FloatLiteral;
 use Serhii\GoodbyeHtml\Ast\Statements\HtmlStatement;
 use Serhii\GoodbyeHtml\Ast\Statements\IfStatement;
 use Serhii\GoodbyeHtml\Ast\InfixExpression;
-use Serhii\GoodbyeHtml\Ast\Literal\IntegerLiteral;
+use Serhii\GoodbyeHtml\Ast\Literals\IntegerLiteral;
 use Serhii\GoodbyeHtml\Ast\Statements\LoopStatement;
-use Serhii\GoodbyeHtml\Ast\Literal\NullLiteral;
+use Serhii\GoodbyeHtml\Ast\Literals\NullLiteral;
 use Serhii\GoodbyeHtml\Ast\PrefixExpression;
 use Serhii\GoodbyeHtml\Ast\Statements\Program;
 use Serhii\GoodbyeHtml\Ast\Statements\Statement;
-use Serhii\GoodbyeHtml\Ast\Literal\StringLiteral;
+use Serhii\GoodbyeHtml\Ast\Literals\StringLiteral;
 use Serhii\GoodbyeHtml\Ast\TernaryExpression;
 use Serhii\GoodbyeHtml\Ast\VariableExpression;
 use Serhii\GoodbyeHtml\Lexer\Lexer;
