+++ src/Evaluator/Evaluator.php
@@ -5,21 +5,21 @@ declare(strict_types=1);
 namespace Serhii\GoodbyeHtml\Evaluator;
 
 use Serhii\GoodbyeHtml\Ast\Statements\BlockStatement;
-use Serhii\GoodbyeHtml\Ast\Literal\BooleanLiteral;
+use Serhii\GoodbyeHtml\Ast\Literals\BooleanLiteral;
 use Serhii\GoodbyeHtml\Ast\Statements\ExpressionStatement;
-use Serhii\GoodbyeHtml\Ast\Literal\FloatLiteral;
+use Serhii\GoodbyeHtml\Ast\Literals\FloatLiteral;
 use Serhii\GoodbyeHtml\Ast\Statements\HtmlStatement;
 use Serhii\GoodbyeHtml\Ast\Statements\IfStatement;
 use Serhii\GoodbyeHtml\Ast\InfixExpression;
-use Serhii\GoodbyeHtml\Ast\Literal\IntegerLiteral;
+use Serhii\GoodbyeHtml\Ast\Literals\IntegerLiteral;
 use Serhii\GoodbyeHtml\Ast\Statements\LoopStatement;
 use Serhii\GoodbyeHtml\Obj\Env;
 use Serhii\GoodbyeHtml\Obj\Obj;
 use Serhii\GoodbyeHtml\Ast\Node;
-use Serhii\GoodbyeHtml\Ast\Literal\NullLiteral;
+use Serhii\GoodbyeHtml\Ast\Literals\NullLiteral;
 use Serhii\GoodbyeHtml\Ast\PrefixExpression;
 use Serhii\GoodbyeHtml\Ast\Statements\Program;
-use Serhii\GoodbyeHtml\Ast\Literal\StringLiteral;
+use Serhii\GoodbyeHtml\Ast\Literals\StringLiteral;
 use Serhii\GoodbyeHtml\Ast\TernaryExpression;
 use Serhii\GoodbyeHtml\Ast\VariableExpression;
 use Serhii\GoodbyeHtml\Obj\BlockObj;
